Refactor: libcrmcommon: move operation-related functions to separate source code file
line count before:
 2030 utils.c

after:
 1456 lib/common/utils.c
  593 lib/common/operations.c
