+++ include/crm_internal.h
@@ -133,14 +133,6 @@ gboolean check_utilization(const char *value);
 long crm_get_sbd_timeout(void);
 gboolean check_sbd_timeout(const char *value);
 
-/* Shared PE/crmd functionality */
-void filter_action_parameters(xmlNode * param_set, const char *version);
-
-/* Resource operation updates */
-xmlNode *create_operation_update(xmlNode * parent, lrmd_event_data_t * event,
-                                 const char * caller_version, int target_rc, const char * node,
-                                 const char * origin, int level);
-
 /* char2score */
 extern int node_score_red;
 extern int node_score_green;
@@ -151,11 +143,17 @@ extern int node_score_infinity;
 int crm_pid_active(long pid, const char *daemon);
 void crm_make_daemon(const char *name, gboolean daemonize, const char *pidfile);
 
+/* from operations.c */
 char *generate_op_key(const char *rsc_id, const char *op_type, int interval);
 char *generate_notify_key(const char *rsc_id, const char *notify_type, const char *op_type);
 char *generate_transition_magic_v202(const char *transition_key, int op_status);
 char *generate_transition_magic(const char *transition_key, int op_status, int op_rc);
 char *generate_transition_key(int action, int transition_id, int target_rc, const char *node);
+void filter_action_parameters(xmlNode *param_set, const char *version);
+xmlNode *create_operation_update(xmlNode *parent, lrmd_event_data_t *event,
+                                 const char *caller_version, int target_rc,
+                                 const char *node, const char *origin,
+                                 int level);
 
 static inline long long
 crm_clear_bit(const char *function, int line, const char *target, long long word, long long bit)
